Support FSx for OpenZFS and NetApp ONTAP in the cluster details' storage section
If we are using a ParallelCluster version with a minor equal or greater than 2,
like 3.2.0b, support showing a link to FSx for OpenZFS and NetApp ONTAP file systems
in the cluster details' storage section.

In the case of ONTAP and OpenZFS we cannot mount an entire file system, so the id
links to the volume details, not to the file system details page. This is similar to
what happens with Ebs, while Efs and FSx for Lustre can mount the entire file system.