Make wizard compliant with limits for external storage attachments
This patch makes the wizard compliant with the limits for external
storage attachments.

When creating a new storage, we check that we are not violating the
constraint of the maximum number of storage of that given type we can
attach. The toggle that allows to switch between creation and attachment
of an existing storage is disabled according to the constraints of the
maximum number of storage that can be created or of existing storage that
can be attached.