Use the official PC API template
This patch makes PCM use the official PC API template, rather than our
copy. We were using the copy because we were missing a change included
in this PR:

https://github.com/aws/aws-parallelcluster/pull/3886/files

in the template that was in the S3 repo. This has now been released, so
we can directly use the template in the S3 repo.