Fixing violation related to Defend Against Reverse Tabnabbing Attack
Reverse tabnabbing is an attack that is possible when a link is
configured to open a new tab in a browser. If done incorrectly,
the new page is able to control the referrer and opener
objects of the parent window, and use that control to replace
the parent site with a phishing site.
The mitigation is add the noopener and noreferrer values in
 the rel attribute in <a> tags.