Remove superfluous SNS topic from PCM Cfn template
This patch removes the unused EcrImageBuilderSNSTopic from the
PCM CloudFormation template. The topic was initially added to
the PC API CloudFromation template in:

- https://github.com/aws/aws-parallelcluster/pull/3054

and was used when updating the Docker image to switch the image
backing the PC API Lambda function. This SNS topic is not used in
the PCM template and can therefore be removed.